Move ConstraintMap-related functions and types into new ConstraintMap.hs file
