+++ code/drasil-code-base/Language/Drasil/Chunk/CodeBase.hs
@@ -9,12 +9,10 @@ import Database.Drasil (ChunkDB, symbResolve)
 
 import Language.Drasil.Code.Expr (CodeExpr)
 import Language.Drasil.Code.Expr.Extract (eDep, eDep')
-import Language.Drasil.Code.Expr.Render (renderConstraint)
 
 import Utils.Drasil.Strings (toPlainName)
 
 import Data.List (nub)
-import qualified Data.Map as Map
 
 -- not using lenses for now
 -- | A 'CodeIdea' must include some code and its name. 
@@ -141,28 +139,3 @@ listToArray c = newSpc (c ^. typ)
           (c ^. term) (getA c) (Array t) (symbol c Implementation) (getUnit c)) 
           Var) (c ^. obv)
         newSpc _ = c
-
--- TODO: Where should this code below belong? Seems odd to be in this file.
-
--- | Type synonym for 'Constraint CodeExpr'
-type ConstraintCE = Constraint CodeExpr
-
--- | Constraints map. Contains all 'Constraint's.
-type ConstraintCEMap = Map.Map UID [ConstraintCE]
-
--- | Creates a map from 'UID' to 'Constraint's for constrained chunks.
-constraintMap :: (HasUID c, Constrained c) => [c] -> ConstraintCEMap
-constraintMap = Map.fromList . map (\x -> (x ^. uid, map renderConstraint $ x ^. constraints))
-
--- | Returns a pair of a chunk and its physical constraints.
-physLookup :: HasUID q => ConstraintCEMap -> q -> (q, [ConstraintCE])
-physLookup m q = constraintLookup q m (filter isPhysC)
-
--- | Returns a pair of a chunk and its software constraints.
-sfwrLookup :: HasUID q => ConstraintCEMap -> q -> (q, [ConstraintCE])
-sfwrLookup m q = constraintLookup q m (filter isSfwrC)
-
--- | Returns a chunk and a filtered list of its constraints.
-constraintLookup :: HasUID q => q -> ConstraintCEMap
-                      -> ([ConstraintCE] -> [ConstraintCE]) -> (q, [ConstraintCE])
-constraintLookup q m filt = (q, maybe [] filt (Map.lookup (q ^. uid) m))
